+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Models/MainFormOrderDetailsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  MainFormOrderDetailsViewModel.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 20.01.2023.
+//
+
+import Foundation
+
+struct MainFormOrderDetailsViewModel {
+    let amountDescription: String
+    let amount: String
+    let orderDescription: String?
+}