+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Presenter/MainFormPresenter.swift
@@ -25,7 +25,28 @@ final class MainFormPresenter {
 // MARK: - IMainFormPresenter
 
 extension MainFormPresenter: IMainFormPresenter {
-    func viewDidLoad() {}
+    func viewDidLoad() {
+        let orderDetails = MainFormOrderDetailsViewModel(
+            amountDescription: "К оплате",
+            amount: "10 500 ₽",
+            orderDescription: "Заказ №123456"
+        )
+
+        let paymentControls = MainFormPaymentControlsViewModel(
+            buttonType: .primary(title: "Оплатить картой")
+        )
+
+        let header = MainFormHeaderViewModel(
+            orderDetails: orderDetails,
+            paymentControls: paymentControls
+        )
+
+        view?.updateHeader(with: header)
+    }
 
     func viewWasClosed() {}
+
+    func viewDidTapPayButton() {
+        router.openCardPaymentForm()
+    }
 }