+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Router/MainFormRouter.swift
@@ -9,4 +9,6 @@ import UIKit
 
 final class MainFormRouter: IMainFormRouter {
     weak var transitionHandler: UIViewController?
+
+    func openCardPaymentForm() {}
 }